refactor: :truck: Move LazyPrec to new file in frontend
