+++ parsley/shared/src/main/scala/parsley/internal/deepembedding/frontend/PrecedenceEmbedding.scala
@@ -3,47 +3,7 @@ package parsley.internal.deepembedding.frontend
 import parsley.internal.deepembedding.backend, backend.StrictParsley
 import parsley.internal.deepembedding.ContOps
 import parsley.internal.deepembedding.ContOps.{result, suspend, ContAdapter}
-import parsley.expr.Fixity
-import parsley.internal.deepembedding.frontend.LazyPrec.Atoms
-import parsley.internal.deepembedding.frontend.LazyPrec.Level
-
-// LazyPrec needs to include an operators table, containing operators (LazyParsleys), their fixity, and precedence
-// It also needs to include an atoms list, with LazyParsley parsers for the atoms
-// final case class LazyOperator[F <: Fixity, A, B](
-//   fixity: F,
-//   precedence: Int,
-//   parser: LazyParsley[F#Op[A, B]]
-// )
-
-// final case class LazyPrec(operators: List[LazyOperator[_, _, _]], atoms: List[LazyParsley[_]])
-
-sealed trait LazyPrec[+Out] {
-  type In
-}
-
-object LazyPrec {
-  final case class Atoms[A](atoms: List[LazyParsley[A]]) extends LazyPrec[A] {
-    type In = A
-  }
-  
-  final case class Level[A, B](
-    lower: LazyPrec[A],
-    fixity: Fixity,
-    operators: List[LazyParsley[Fixity#Op[A, B]]]
-  ) extends LazyPrec[B] {
-    type In = A
-  }
-
-  def fromPrec[A](table: parsley.expr.Prec[A]): LazyPrec[A] = {
-    import parsley.expr.{Atoms => PAtoms, Level => PLevel}
-
-    table match {
-      case PAtoms(atom0, atoms @ _*) => Atoms((atom0 +: atoms).toList.map(_.internal))
-      case PLevel(lower, ops) => Level(fromPrec(lower), ops.f, ops.operators.toList.map(_.internal))
-    }
-  }
-}
-
+import parsley.internal.deepembedding.frontend.LazyPrec.{Atoms, Level, traverse_ => traversePrec_}
 
 object TraverseHelper {
   def traverse[M[_, +_]: ContOps, R, A, B](xs: List[A])(f: A => M[R, B]): M[R, List[B]] = xs match {
@@ -57,18 +17,6 @@ object TraverseHelper {
   def traverse_[M[_, +_]: ContOps, R, A, B](xs: List[A])(f: A => M[R, B]): M[R, Unit] = xs match {
     case Nil => result(())
     case x :: xs => f(x) >> traverse_(xs)(f)
-  }
-}
-
-object LazyPrecTraverseHelper {
-  def traverse_[M[_, +_]: ContOps, R, B](
-    table: LazyPrec[_]
-  )(
-    fAtom: LazyParsley[_] => M[R, B],
-    fOp: LazyParsley[_] => M[R, B]
-  ): M[R, Unit] = table match {
-    case Atoms(atoms) => TraverseHelper.traverse_[M, R, LazyParsley[_], B](atoms)(fAtom)
-    case Level(lower, _, ops) => traverse_[M, R, B](lower)(fAtom, fOp) >> TraverseHelper.traverse_[M, R, LazyParsley[_], B](ops)(fOp)
   }
 }
 
@@ -95,7 +43,7 @@ private [parsley] final class Precedence[A](table: LazyPrec[A]) extends LazyPars
   // }
 
   override protected def findLetsAux[M[_, +_]: ContOps, R](seen: Set[LazyParsley[_]])(implicit state: LetFinderState): M[R,Unit] = {
-    LazyPrecTraverseHelper.traverse_(
+    traversePrec_(
       table
     )(
       a => suspend[M, R, Unit](a.findLets(seen)),
